Move application annotation and path to the application.properties to allow automated rest assured configuration.
Replace more deprecated mutiny methods with their new naming.

Add more extensive testing to the REST interface and add validation of
id to be UUID as to not cause an error 500 if something else is
supplied.

Allow configuration of the MockServerClient from any test with more
options (and a couple of helpers).