NOTIF-192 Silence successful health checks logs (#285)
* NOTIF-192 Silence successful health checks logs

Health checks produce logs lines that are unnecessary so we want to get rid of these kind of logs.

* NOTIF-192 Silence successful health checks logs

Health checks produce logs lines that are unnecessary so we want to get rid of these kind of logs.

* Remove JSONB dependency

Remove JSONB dependency as it leats to strange exceptions. Seems like it's somehow not that nicely working together with the jackson dependency.