refactor: include outgoing headers in the header factory
This might help us in the future if we copy this class to implement
something similar for another service. Not having those headers copied
might end up in some headaches for the developer, wondering why it
doesn't work. So including it here doesn't hurt, and might benefit us in
the future.

RHCLOUD-22065