[RHCLOUD-22178] Include environment URLs when serializing Actions (#1567)
* refactor: remove the "transform" function in favor of "toJsonObject"

Instead of having a function that simply calls the "toJsonObject", I
thought that we might as well have that function only.

NOTIF-823

* feature: test for the "toJsonObject" function

Basically adds a test to ensure that we don't break the function's
implementation.

NOTIF-823

* feature: include the environment's URL on sent OpenBridge events

This way every action comes associated with the URL it can be acted
upon, or checked on, to avoid any confussion.

RHCLOUD-22178

* feature: test all the values for the Open Bridge event

This way we can be sure that whenever we change something when sending
events to OpenBridge, we are doing it on purpose, and not by mistake.

RHCLOUD-22178

* refactor: rename OpenBridge to RHOSE

The product was officially renamed time ago, and we should be using RHOSE
instead.

RHCLOUD-22178

* refactor: remove redundant tests

These tests are not needed since they are already covered in the new
test.

RHCLOUD-22178

* refactor: move constants to main classes

Having these constants in the main classes will allow us to reuse them
wherever we want.

RHCLOUD-22545

* refactor: typo

RHCLOUD-22545

* refactor: remove key checks from primitives

These primitives will be tested by their values, so if any key for them
is missing, the "test the value" tests should signal that.

RHCLOUD-22545

* refactor: make use of the expected data instead of the constants

In this test we are testing if the transformation went well. Therefore,
it makes sense to use the original object to make the comparisons,
instead of using the constants.

RHCLOUD-22178

* refactor: remove redundant key checks for the JSON object

Event though they might be useful, they're redundant, because we are
also checking the values for them below. This could lead to duplicated
errors, so we might as well just keep the value checks, which should
cover both the key and the value itself.

RHCLOUD-22178

* refactor: put the "environment_url" inside the "data" object

Gwenneg said that otherwise this might not work, so in order to be safe
I'm just moving that key as suggested.

RHCLOUD-22178