refactor: reformat the POMs using the Quarkus CLI
This will standardize the POMs and make it easier to review the changes
when we use the CLI to add or remove dependencies. The command used to
reformat the POMs was the following one:

* quarkus extension remove quarkus-rest-client-mutiny

None of the POMs had that extension, so that way I made sure that the
command was basically a NOP that reformatted the files.

RHCLOUD-22602