[RHCLOUD-22065] Feature store webhook secrets sources (#1561)
* feature: database migration and model for Sources support

In order to store the external IDs of the Sources authentications, we
need to create a couple of columns on the "Camel" and "Webhook"
properties' tables.

RHCLOUD-22065

* feature: interface for common secret fields

Both the "Camel" and the "Webhook" properties share the "basic
authentication" and "secret token" fields, but in order to access them
in an easier way without having to rely on using "instanceof" every time
we need to access those properties, an interface is used instead of
using inheritance.

RHCLOUD-22065

* feature: Sources REST client and utilities to manage secrets

Implements the requried REST client and the utility class which holds
some logic in order to manage secrets in Sources.

RHCLOUD-22065

* feature: feature flag for using Sources as the secrets backend

Using a feature flag will provide us with the ability of switching the
Sources backend on and off, at will.

RHCLOUD-22065

* feature: implement Sources as the secret back end with feature flags

Makes the hanlders manage Sources secrets if the feature flags allow to
do so.

RHCLOUD-22065

* feature: validate the basic authentication object

This makes sure that if the object isn't null, the user must specify
non-blank values for the password and the username of the object.

RHCLOUD-22065

* feature: log important events

A few log messages have been added to the important operations in order
to help in the debugging of the potential issues in the future.

RHCLOUD-22065

* refactor: merge the "add column" statements in one

Makes the statements more concise by achieving the same thing.

RHCLOUD-22065

* fix: secrets are not being udpated

The problem was that I was grabbing the "endpoint" object sent by the
user, which doesn't contain any IDs for the Sources secrets, instead of
the database object. The latter is the one which contains the required
information to send the update requests to Sources.

RHCLOUD-22065

* refactor: use wrapped long instead of the primitive for secret IDs

We might have "nulls" in the database, so it makes it easier to handle
that case if we are using wrapped values.

RHCLOUD-22065

* refactor: rename "get secrets" function name

It makes it more accurate.

RHCLOUD-22065

* refactor: remove redundant "@ApplicationScoped" annotation

Quarkus already adds that scope automatically.

RHCLOUD-22065

* refactor: use "RestPath" instead of "PathParam" in the RESTClient

The annotation is more developer friendly and saves us from specifying
the path parameter, which is kind of redundant.

RHCLOUD-22065

* refactor: typo on the comment

RHCLOUD-22065

* refactor: ignore incoming unknown properties for safety

This will prevent us from having errors if Sources sends a payload with
properties we haven't got defined in our class.

RHCLOUD-22065

* refactor: check the instance of the endpoint properties before updating

The code could throw a "ClassCastException" otherwise.

RHCLOUD-22065

* refactor: point to Swagger UI for the OpenAPI specifications

Instead of pointing at the RAW JSON file which is harder to parse, we
can simply point the user at the Swagger UI that Red Hat provides.

RHCLOUD-22065

* refactor: include outgoing headers in the header factory

This might help us in the future if we copy this class to implement
something similar for another service. Not having those headers copied
might end up in some headaches for the developer, wondering why it
doesn't work. So including it here doesn't hurt, and might benefit us in
the future.

RHCLOUD-22065

* feature: implement different behaviours for the secrets update operation

There were more cases which Gwenneg brought up and made total sense:

- If the user wants to update an endpoint and remove its secrets, they
  probably will send the "basic authentication" and "secret token"
  secrets empty, which should trigger the deletion.
- On the other hand, if they didn't create the secrets initially, and
  they want to create them now by updating the endpoint, the will
  probably send the those two secrets along with the updated endpoint.

Therefore, the update operation should cover those cases too.

RHCLOUD-22065

* refactor: correct the configuration path for Sources API

Clowder will put the URL of the Sources service on the "endpoints.soures-api"
endpoint, so we should be reading the URL from that key.

RHCLOUD-22065

* feature: sources as a Clowder dependency

In order for the backend to work, we need to set Sources as a Clowder
dependency.

RHCLOUD-22065

* feature: fault tolerance for Sources

The few retries will help us ensure that when the REST service returns
an error, it's not a "one-off" kind of error, but rather something that
should be looked at.

RHCLOUD-22065

* refactor: remove validation for the "Basic Authentication" object

This could cause issues with the front end[1], so we might want to hold
this back until we're safe to implement these validations.

[1]: https://github.com/RedHatInsights/notifications-backend/pull/1561#discussion_r995458415

RHCLOUD-22065

* refactor: rename the database migration

Another migration was added while the PR associated to this branch was
open.

RHCLOUD-22065

* refactor: move fault tolerance dependency to "backend" module

This way it only gets added to the module that needs it.

RHCLOUD-22065

* refactor: specify the timeout via properties

Improves the flexibility to change it as needed via environment
variables.

RHCLOUD-22065

* refactor: use a "WebApplicationException" instead of a runtime one

Plays nicer with Quarkus.

RHCLOUD-22065

* refactor: remove redundant condition

Whenever the secret IDs are missing from the model, the "Long" variable
will always be null. Probably these conditions were left out when
primitives were being used instead.

RHCLOUD-22065

* refactor: NOOP when the secret token is blank

In this case if the user just sends a blank secret token we will also
assume that they want it deleted or that they made a mistake.

RHCLOUD-22065

* refactor: finish the unfinished comment

Apparently I forgot to finish writing the comment.

RHCLOUD-22065

* refactor: remove redundant comment in the code

Basically it was explaining the same thing that is explained in the
Javadoc of the method.

RHCLOUD-22065

* refactor: NOOP when the basic authentication's fields are blank

Similar to the NOOP we do when the secret token is blank, we should also
make sure that we don't hit sources when we receive blank basic
authentication properties.

RHCLOUD-22065

* refactor: NOOP when the "secret token" is blank

Avoids creating empty secrets in Sources.

RHCLOUD-22065

* refactor: use "verifyNoInteractions" in Mockito

Makes the code easier to follow.

RHCLOUD-22065

* refactor: move dependency to the "Quarkus" section

I had put it with the "test" dependencies, and it didn't make much sense
to have it there.

RHCLOUD-22065

* refactor: add "todo" comment

We might end up in a situation of having dangling secrets in Sources if
the endpoint creation fails. The "todo" comment mentions this and links
to a Jira ticket.

RHCLOUD-22065