refactor: move the "SecretUtils" to the "common" module
They need to be used in the "engine" module too, so in order to avoid
repeating code, the move had to be done. The tests, though, are kept in
the "backend" module, since they seem to trigger a database connection
requirement, and that's just too much for the "common" module.

RHCLOUD-22065