[RHCLOUD-22065] Feature sources processors (#1604)
* refactor: move the "SecretUtils" to the "common" module

They need to be used in the "engine" module too, so in order to avoid
repeating code, the move had to be done. The tests, though, are kept in
the "backend" module, since they seem to trigger a database connection
requirement, and that's just too much for the "common" module.

RHCLOUD-22065

* feature: call Sources in the endpoint processors

The idea is that if the Sources integration is enabled, then the
processor will call Sources to fetch the secrets. If not, the secrets
are pulled from the database: the thing we have been doing until now.

RHCLOUD-22065

* fix: tests in the backend and the common module

Some tests were not running because they were missing the "Sources REST
client URL" property when running the tests. By setting up a test
profile for the common module and setting up the property in the
backend's test properties the issue seems to be gone.

RHCLOUD-22065

* fix: Sources service name

The name is "sources-api-svc", not "sources-api".

RHCLOUD-22065

* refactor: use "pgsql" instead of "h2" for the tests

Since the database is not really used, I initially thought of using "h2"
to make Quarkus happy. However Gwenneg pointed out that it might be
better to just use "pgsql" for consistency.

RHCLOUD-22065

* refactor: move the "SecretUtilsTest" back to the backend

For now we are going to keep it in the backend module, because there's a
tech debt story about moving commont tests to the "common" module, but
everything wants to be tackled at once.

RHCLOUD-22065

* refactor: remove unused dependencies

RHCLOUD-22065