refactor: move the test action creation to its own function
The responsability of the "sendTestEvent" function should be to send
that test event, not to both create it and send it. By moving the
creation of the test action to a new class, reusability is improved, and
it will make it easier to generate test actions in tests if they're
required.

RHCLOUD-22470