Add a lot more information to each rule
The mock-up of the output lists the description, risk, incident flag, and
number of affected systems for each rule.  We need to pull this from the
rule information when we create a new entry in the map for the new,
resolved and deactivated recommendations.

We're also going to simplify this by making a major assumption.

Let us assume that in general a system will only have a new or resolved
event for each recommendation once per day.  Systems upload once per
day in general, and each upload can never report the same rule as new or
resolved twice (because each rule impacts a system only once).  The only
situation in which this could be false would be for a system to upload
at least three times - once to report a new rule, then to resolve it,
and finally to report the same rule as new again.  That's got to be
pretty rare, right?

With that assumption, we do not have to remember the set of Inventory
IDs reporting each rule as new or resolved; we just have to increment
a count each time a new event is seen.

Signed-off-by: Paul Wayper <paulway@redhat.com>