feature: limit the maximum allowed number of behavior group creations
QE has spotted that after a few thousand behavior groups that get
created, the frontend's performance degrades. However, having a customer
with more than 15 behavior groups configured is going to be a rare case.

That is why we are limiting the number of behavior groups that can be
created, since we assume that if more need to be created, there will
highly likely be old ones that are not used, and therefore need to be
deleted first.

RHCLOUD-21842