refactor: use a REST call from the backend to trigger a test event
Kafka should not be used in the backend module because we want to avoid
the service getting stuck or going down if we face a Kafka failure or
Kafka lag. Therefore, in order to communicate a "test endoint
integration" request from the client, we need to use a REST call from
the backend to the engine.

In the engine, an event is sent to the Kafka queue, because that way it
will get processed through the whole "EventConsumer" flow.

RHCLOUD-22470