refactor: emit a string instead of an action for the event
Quarkus was having some issues to serialize and emit the test event.
Also, the event consumer expects a string message, so I thought of
simply using the string class for emitting the message, which solved the
issue.

The tests have been updated to reflect the needed changes for everything
to work as it does in runtime.

RHCLOUD-22470