fix: Sources not receiving the "x-rh-identity header" from the engine
Sources was returning 401 responses when the engine was trying to fetch
the related secrets to an endpoint. This was happening because the
Sources REST Service class was configured to forward any "x-rh-identity"
header received from the client, but in the case of the engine, since
there was no header to forward, the request was being sent without any
header.

By switching to an explicit header specification for the REST Service
class, we force the caller to provide the proper "x-rh-identity"
header's content. Also, the secret utilities have been updated to honor
this interface.

Finally, the tests have been updated as well to make sure that the
secret utilities are sending the headers in every single case.

RHCLOUD-23722