feature: implement PSK authentication with Sources (#1721)
The initial implementation simply forwarded the received x-rh-identity header
to Sources. This worked for the backend, but not the engine, since the engine
doesn't have a client that sends such header.

The second implementation generated a minimal x-rh-identity header with
an org-id inside of it, but it also didn't work because RBAC expects an
username to identify the user. This also had the drawback of the user needing
the Sources administrator role.

This implementation uses the PSK to authenticate to Sources, and then a
second header so that Sources knows to which tenants the operations tie
to.

RHCLOUD-23966