feature: checked exceptions for the "SecretUtils" service
The checked exceptions will force the caller to explicitly deal with any
errors that may occur during a Sources request call. The goal of this is
to force the developer to think about which actions should be taken when
the external service is reporting errors.

The reason for this change is that the "get endpoints" endpoint was
reporting 404 errors when at least one endpoint had references to
non-existing Sources secrets. In theory this should not happen, because
we already took good care of making sure that every operation leaves the
Sources references as they should in the RHCLOUD-24435[1] ticket.

However, there are a couple of "unlikely" scenarios in which this may
happen:

* The secrets dissapear in Sources for any unexpected reason. One of
  them could be, for example, a bad data migration on their side.
* QE tests may spawn two Sources instances, and this could make some
  tests fail if they end up creating secrets with cross references to
  those instances.

[1]: https://issues.redhat.com/browse/RHCLOUD-24435

[RHCLOUD-25078]