feature: pull events from the database with date filters
Events will be the first entities that we will export when clients
request it. The idea is to mimic what a user would see in the "event
log", so that is why just a subset of the columns and rows are pulled
from the database, instead of pulling the whole rows.

RHCLOUD-25597