fix: avoid exceptions blocking the migrations
During the migration, we saw that constraint violation exceptions on
endpoints with invalid URLs were rolling back the whole operation. The
goal of this change is to handle those exceptions in order to be able to
continue with the process even when a failure happens, and in order to
handle the deletion of any created secrets when the error occurred.

RHCLOUD-24734