[RHCLOUD-25597] Export service integration — Request handler (#1914)
* feature: handle the export requests

Handles the incoming export requests accordingly. The handler makes sure
that the export requests are requested for Notifications for the
supported resources and formats. It also validates that the incoming
inputs are supported and valid, before querying the database and
transforming the results.

RHCLOUD-25597

* refactor: add the "test launch mode" check for runtime overrides

RHCLOUD-25597

* refactor: use the same group ID for Kafka as the rest

RHCLOUD-25597

* refactor: rename "exports" service to "export" service

RHCLOUD-25597

* refactor: extract the common path to a class annotation

RHCLOUD-25597

* refactor: remove unnecessary transactional annotation

RHCLOUD-25597

* refactor: use a different switch for the integration

We can use the switch of the Kafka consumer to control the integration,
and even better, we will be able to NOT consume and discard any messages
from the topic until we decide so.

RHCLOUD-25597

* refactor: simplify Kafka message consumption

We are not going to access the Kafka headers, so there is no need for
the message wrapper.

RHCLOUD-25597

* refactor: log the stack trace for more information

When errors occur, we will like to see as much as information as
possible, and the stack trace is an important piece of that information.

RHCLOUD-25597

* refactor: use the "application" constant where due

RHCLOUD-25597

* refactor: move classes to more accurate packages

The exception class and the interface just make more sense to be in the
transformers package.

* refactor: separate the filters extraction to a different class

This should reduce the responsabilities of the handler, and make future
filters easier to extract.

RHCLOUD-25597

* fix: typo

RHCLOUD-25597

* refactor: set the export service as an optional dependency

RHCLOUD-25597

* refactor: make logs more readable

RHCLOUD-25597

* refactor: extract failures counter to a variable

RHCLOUD-25597

* refactor: make the transformations return when failing

Otherwise the successes counter would be incremented.

RHCLOUD-25597

* refactor: reduce the number of required arguments for the function

The arguments were required for debug logging, so it feels a bit
unnecessary to have so many arguments in the signature. Plus, after the
schema change that is coming, the export request's UUID will come in the
payload and will be accessible again.

RHCLOUD-25597

* refactor: extract the events export code to a service

This makes the code way easier to follow and makes a clear separation of
responsibilities.

RHCLOUD-25597

* refactor: make sure only the correct counters are incremented

RHCLOUD-25597

* refactor: move the "successes counter" to a variable

RHCLOUD-25597

* refactor: return on error to avoid incrementing the successes counter

RHCLOUD-25597

* refactor: surround the handler with a try-catch block

Avoids Kafka poison pills and catches any exceptions when calling the
export-service.

RHCLOUD-25597

* refactor: use the correct UUIDs in the tests

We were using the same UUID for the export request's UUID and the
resource request's UUID. The former is actually the parent of the
latter, so we needed to make that distinction when building the test
Cloud Event too.

RHCLOUD-25597

* refactor: use mocks for testing the handler

Using mocked services is just easier for testing the handler's logic. On
the other hand, since we needed to test that the correct "Content-Type"
headers are being sent, we keeped the Mock Server for testing those very
specific cases.

RHCLOUD-25597

* feature: Kafka consumer limits

This should help us better tweak the consumer in case it starts failing
for any reason.

RHCLOUD-25597

* refactor: use the new schema for the export service

Introduces a change where it makes the export request's UUID explicit.

RHCLOUD-25597

* refactor: increment failure metrics on unexpected responses

RHCLOUD-25597

* refactor: typo in environment variable's name

RHCLOUD-25597

* refactor: remove unused injection

RHCLOUD-25597

* feature: reference to the export-service PSK secret

That reference is needed to be able to communicate to the
export-service.

RHCLOUD-25597