fix: compile sources before analyzing them
The scanner was complaining that the ".java" files were not compiled,
and that should be provided via a property. It was working locally
because I always have the compiled project there, so by copying all the
contents of the project's directory it also copied the compiled classes.
However, in a pipeline, where everything is fresh, that didn't happen,
which I suppose is causing the scanner to complain.

RHCLOUD-22207