[RHCLOUD-24734] fix: avoid exceptions blocking the migrations (#1922)
* fix: avoid exceptions blocking the migrations

During the migration, we saw that constraint violation exceptions on
endpoints with invalid URLs were rolling back the whole operation. The
goal of this change is to handle those exceptions in order to be able to
continue with the process even when a failure happens, and in order to
handle the deletion of any created secrets when the error occurred.

RHCLOUD-24734

* refactor: only delete secrets when a Constraint Violation occurs

The rest of the exceptions mean that no secrets were created due to
errors. The constraint violation exception will be thrown when saving
the endpoint, thus, after creating the secrets in Sources.

RHCLOUD-24734