refactor: MBOP URL forced to be set by an environment variable
The URL for the MBOP service cannot be picked up from the Clowder
configuration file, because MBOP is not a Clowderized application.
Therefore, the idea behind this change is to allow it to work locally by
giving it a value, but forcing the developer to actually set a value in
stage or production in Clowder. That way we will not forget that the
parameter must be set.

RHCLOUD-23645