RHCLOUD-23645 | feature: fetch users from MBOP instead of the IT service (#2053)
* feature: fetch users from MBOP instead of the IT service

Notifications shouldn't be directly calling the IT service for users,
since MBOP has been designed and developed to act as the middleman, thus
saving us from having to deal with things like certificates in our
application. Therefore, this is the first step to replace that
integration.

RHCLOUD-23645

* refactor: verify that asking for more pages works as expected

RHCLOUD-23645

* refacor: verify that "offset" arguments are correctly calculated

RHCLOUD-23645

* refactor: MBOP URL forced to be set by an environment variable

The URL for the MBOP service cannot be picked up from the Clowder
configuration file, because MBOP is not a Clowderized application.
Therefore, the idea behind this change is to allow it to work locally by
giving it a value, but forcing the developer to actually set a value in
stage or production in Clowder. That way we will not forget that the
parameter must be set.

RHCLOUD-23645

* refactor: reuse the service's authentication parameters

RHCLOUD-23645

* refactor: remove non-optimal annotation

RHCLOIUD-23645

* refactor: reuse property

RHCLOUD-23645