[RHCLOUD-25186] feature allow overriding default message test endpoint (#1890)
* refactor: split "endpoint test" requests in two

We need to take care of the potential incoming test message that the
user may send, so in order to have a proper name for both of the DTOs, I
decided to rename the old one and use the old name for the new DTO.

RHCLOUD-25186

* feature: allow sending custom messages when testing an endpoint

The second step on sending "test events" for integrations was to allow
sending custom messages on those events. This implements it in a way
that the client can decide whether to send a message or not. But if they
do, the message must not be blank.

RHCLOUD-25186

* refactor: simplify custom message validation

RHCLOUD-25186

* refactor: change reference to constant

RHCLOUD-25186

* Apply review suggestions

* Fix test

---------

Co-authored-by: Gwenneg Lepage <glepage@redhat.com>