fix: the key store location as a default value
The ephemeral environment is complaining about not having the
environment variable set, and therefore not having a proper value to
boot Quarkus with it. Since we already provide a value in the properties
file, it might as well be worth specifying it as a default value in the
configuration file.

RHCLOUD-28170