fix: revert default value and add one for the engine (#2147)
The default value seems to be overridden by Clowder even though we have
the default configuration value in the code. I am reverting the previous
change and adding the environment variable in AppInterface.

Also, the engine is suffering from the same error because no default
value is provided for the connector feature flag. I am providing one
just in case.

RHCLOUD-28170