fix: backoffice proxy's path
The backoffice proxy is returning a 404 at the moment because the path
of the endpoint to send emails is wrong.

The reason for using those environment variable names which don't have
the "NOTIFICATIONS_CONNECTOR" prefix is that they get set in
AppInterface, and therefore we must honor that naming so that they get
properly replaced when the connector gets deployed to stage or
production.

RHCLOUD-28244