feature: cache the extracted usernames from IT and RBAC calls
Calling to IT and RBAC services is a really expensive operation, and
usually the responses are quite static. It makes sense to use caching
for these usernames in order to avoid having to call for that data each
time we need to send an email, specially when there is a huge batch of
emails to be sent.

RHCLOUD-28138