RHCLOUD-28631 | Fix: connector sends duplicate emails (#2230)
* fix: connector sending duplicate emails

The connector was sending duplicate emails because we were processing
the recipient settings in a for loop and sending an email for each one
of them.

However, the legacy implementation reduced all the fetched users from
the different user providers into a single list, which the email
connector didn't do.

The strategy followed for the fix is to split the exchanges and then
aggregate them, which also aggregates the users, producing a single
exchange to take care of, and therefore, a single email if the
configuration option is enabled. If not, we still can send one email per
user, even though that option is deprecated.

RHCLOUD-28631

* refactor: remove unused dispatcher

RHCLOUD-28631

* refactor: set group UUID before executing the group route

The dispatcher processor took care of that, and since we are doing it in
a route now, we need to set it up manually again.

RHCLOUD-28631

* refactor: set usernames property before the split

RHCLOUD-28631

* refactor: send only one success exchange to Kafka

Until now we were sending one success exchange per email sent, which
isn't correct. We should only be sending one success exchange per whole
operation.

RHCLOUD-28631