feature: tests for the endpoints and service implementation
We are making an effort to stop returning secrets in responses, and at
the same time, signal if they are present or not, so that the UI can
effectively show feedback about that.

Since we need a transition period until the UI moves to the new way of
managing secrets, we still support both ways of managing them.

The tests are in a separate commit and branch to make reviewing easier.

RHCLOUD-27431