fix: column names not being read due to the spaces
The Business Insights team cannot properly index the columns which have
spaces on them, so the goal of these changes is to replace them with
underscores. And since we're at it, in order to make it consistent
everywhere and with every query, we are also naming all the columns in
lowercase.

RHCLOUD-31387