fix: endpoint subentities not properly serialized
I removed all the annotations for the internal entities, because I
assumed we had all the bases covered when I changed the response entity
from "Endpoint" to "EndpointDTO" in all of the REST endpoints that we
have.

However, there are some other endpoints that return some other entities
like behavior groups, which also serialize endpoints. But they serialize
the internal model's endpoint, not the DTO model's one. And since we are
changing our model slowly by slowly, we decided not to create DTOs for
our entire internal model in one single step, since it would be too much
work and very hard to review.

Gwenneg suggested bringing the JSON serialization annotations back,
and all the tests passed everywhere, but we faced another problem in
the UI: it was breaking in certain calls where the endpoints where
being listed.

I realized that even though I reverted the "Endpoint"'s entity's JSON
serialization annotations, I did not do it for the "Endpoint"'s
properties. That is what probably is causing the issue with the UI,
since the properties are not being properly serialized.

Finally, I'm reverting the functions that change the case of the
"EndpointType"'s enum values, because we are supposed to return the
values in snake case, and that was not working for the
"EMAIL_SUBSCRIPTION" enum value.

RHCLOUD-31165