feature: gracefully handle gRPC exceptions
When gRPC throws exceptions Quarkus simply returns the entire body of
the exception in the response to the client. It usually it is not useful
because it's hard to follow, and the client should not really receive
this kind of information.

This change logs the exception with all kinds of information and returns
an "internal server error" response to the client omitting the details
of what caused the error.

RHCLOUD-32502