refactor: move test for fetching payloads from the engine
The main problem is that having a generic test was not working due to
every connector needing a special payload checker. The goal of the test
is to make sure that the common route calls the engine when required, so
when attempting to mock the end of the common route it would make the
rest of the tests fail.

That is why it is moving to the engine connector. Just because we might
change our connectors, but email will probably stay there for way
longer.

RHCLOUD-33142