feature: migrate duplicated endpoint and behavior groups' names
The problem with using the migration endpoint for this is that we don't
know when the Fedramp environment will be avaialble to run that
migration. And if we push the database constraints before the migration
has run in Fedramp, Notifications might crash. So the way to ensure that
this does not happen is by running the data migrations as a database
migration.

RHCLOUD-33636