RHCLOUD-33636 | feature: migrate duplicated endpoint and behavior groups' names (#2807)
* feature: migrate duplicated endpoint and behavior groups' names

The problem with using the migration endpoint for this is that we don't
know when the Fedramp environment will be avaialble to run that
migration. And if we push the database constraints before the migration
has run in Fedramp, Notifications might crash. So the way to ensure that
this does not happen is by running the data migrations as a database
migration.

RHCLOUD-33636

* feature: forbid duplicate behavior group display names

Prevents creating behavior groups with duplicate display names in the
same organization.

RHCLOUD-33636

* feature: forbid duplicate integration names

Prevents creating integrations with duplicate names in the same
organization.

RHCLOUD-33635

* feature: remove duplicate name migration code

Once the migration has run in production we have to remove the code
which is related to it so that we can merge the database constraints.
If we try to merge the database constraints first the tests for the
migration will rightfully fail.

RHCLOUD-33636