RHCLOUD-33142 | feature: store heavy payloads in the database for the connectors (#2763)
* feature: retrieve the payloads from the engine

Some payloads are way too heavy to be sent via Kafka. In order to avoid
that problem, the engine will expose a REST endpoint which will allow
the connectors to fetch the event's payloads via that endpoint.

RHCLOUD-33142

* refactor: move test for fetching payloads from the engine

The main problem is that having a generic test was not working due to
every connector needing a special payload checker. The goal of the test
is to make sure that the common route calls the engine when required, so
when attempting to mock the end of the common route it would make the
rest of the tests fail.

That is why it is moving to the engine connector. Just because we might
change our connectors, but email will probably stay there for way
longer.

RHCLOUD-33142

* refactor: use the real default maximum Kafka message size value

RHCLOUD-33142

* fix: configuration mock is making paylods to be stored in the database

When mocking the engine's configuration in other tests, the default
value for the maximum Kafka message size is the default value for an
integer in Java: 0. That is making those payloads to be stored in the
database and therefore, making the assertions in other tests fail.

RHCLOUD-33142

* refactor: use a payload ID instead of the event ID for the details

The same event ID might trigger multiple notifications with different
payloads, so we must have a unique ID per payload in the database.

RHCLOUD-33142

* refactor: notifications' engine as mandatory dependency

The connectors needs to be able to contact the engine if they need to
fetch payloads from the engine.

RHCLOUD-33142

* refactor: send the payload identifiers in the Cloud Event

It simplifies the implementation and makes it easier to follow.

RHCLOUD-33142

* refactor: delegate payload deletion to the cascade statement

Instead of manually deleting the payloads from the database, we have
decided to rely on the "ON DELETE CASCADE" statement that the payloads'
table has. That way, whenever we clean up the events the associated
payloads will be cleaned up too.

RHCLOUD-33142

* refactor: remove not needed mock

RHCLOUD-33142

* Update engine/src/test/java/com/redhat/cloud/notifications/routers/PayloadDetailsResourceTest.java

Co-authored-by: Guillaume Duval <117720964+g-duval@users.noreply.github.com>

---------

Co-authored-by: Guillaume Duval <117720964+g-duval@users.noreply.github.com>