feature: incrase transaction timeouts for the bundles operations
We are suffering from timeouts in production when we attempt to modify
or delete an existing bundle. This is due to the "events" table being
huge, and the "bundle_id" column not being indexed.

Adding an index to that column for this purpose seems overkill for now,
especially because these endpoints don't get used very often.

RHCLOUD-34047