refactor: correct the high volume Kafka topic names
We have seen that things don't break if the topic name is not correct.
The email connector shows lots of warnings, but it keeps processing the
events. The engine does not complain about it.

This means that until the high volume topic is enabled in other
environments, we should not suffer from errors or crashloops.

RHCLOUD-34317