feature: log when no integration IDs have been returned
When the subject is not authorized to fetch any of the workspace's
integrations or when there are none to return, we simply respond with an
empty integrations page for the user.

Since Kessel is a fairly new integration, we think that we need to log
this behavior so that we can debug and understand it when it happens at
these initial stages.

RHCLOUD-32983